reorganization of the structure of files and folders according to requirements
